+++ backend-java/src/main/java/ru/savefood/shop/dto/ReceiptLotDraft.java
@@ -3,6 +3,6 @@ package ru.savefood.shop.dto;
 /** One lot draft grouped from receipt items (schemas.py ReceiptLotDraft). */
 public record ReceiptLotDraft(
     String description,
-    Integer quantity,
+    Double quantity,
     String category) {
 }
